black/flake/isort formatting, minor documentation details

+++ sportsdataverse/mlb/mlb_loaders.py
@@ -1,50 +1,39 @@
-import json
-import pandas as pd
-from pandas import json_normalize
-from re import search
-from datetime import datetime
 from sportsdataverse.dl_utils import download
 
 import os
-def mlbam_copyright_info(saveFile=False,returnFile=False):
-	"""
-	Displays the copyright info for the MLBAM API.
-
-	Args:
-		saveFile (boolean) = False
-		If saveFile is set to True, the copyright file generated is saved.
-
-		returnFile (boolean) = False
-		If returnFile is set to True, the copyright file is returned.
-
-	"""
-	url = "http://gdx.mlb.com/components/copyright.txt"
-	resp = download(url=url)
-	try:
-		if resp is not None:
-			l_string = str(resp, 'UTF-8')
-			with open("mlbam_copyright.txt","w+" ,encoding = "utf-8") as file:
-				file.writelines(str(l_string))
-
-			with open("mlbam_copyright.txt", "r" ,encoding = "utf-8") as file:
-				mlbam = file.read()
-
-			if saveFile == False:
-				if os.path.exists("mlbam_copyright.txt"):
-					os.remove("mlbam_copyright.txt")
-				else:
-					pass
-			else:
-				pass
-			print(mlbam)
-
-			if returnFile == True:
-				return mlbam
-			else:
-				pass
 
 
-		else:
-			print('Could not connect to the internet. \nPlease fix this issue to be able to use this package.')
-	except:
-		print('Could not connect to the internet. \nPlease fix this issue to be able to use this package.')
+def mlbam_copyright_info(saveFile=False, returnFile=False, **kwargs):
+    """
+    Displays the copyright info for the MLBAM API.
+
+    Args:
+            saveFile (boolean) = False
+            If saveFile is set to True, the copyright file generated is saved.
+
+            returnFile (boolean) = False
+            If returnFile is set to True, the copyright file is returned.
+
+    """
+    url = "http://gdx.mlb.com/components/copyright.txt"
+    resp = download(url=url, **kwargs)
+    try:
+        if resp is not None:
+            l_string = str(resp.text(), "UTF-8")
+            with open("mlbam_copyright.txt", "w+", encoding="utf-8") as file:
+                file.writelines(l_string)
+
+            with open("mlbam_copyright.txt", "r", encoding="utf-8") as file:
+                mlbam = file.read()
+
+            if saveFile is False and os.path.exists("mlbam_copyright.txt"):
+                os.remove("mlbam_copyright.txt")
+            print(mlbam)
+
+            if returnFile is True:
+                return mlbam
+
+        else:
+            print("Could not connect to the internet. \nPlease fix this issue to be able to use this package.")
+    except Exception as e:
+        print("Could not connect to the internet. \nPlease fix this issue to be able to use this package.", e)
